+++ src/podcast_analyzer/__init__.py
@@ -7,7 +7,7 @@
 
 """Django Podcast Analyzer"""
 
-__version__ = "0.2.8"
+__version__ = "0.3.0"
 
 from podcast_analyzer.exceptions import (
     FeedFetchError,